Update README to README.md in the build files
Since we removed the README file in
https://github.com/google/cmockery/pull/44
the build is broken since it explicitly checks for the existence of a
file literally named README. Also, several make rules also depend on
that same file.

This updates the check to use README.md in place of README everywhere
that configure and Makefile expect it to be.